+++ package/rules/index.js
@@ -4,12 +4,14 @@ const css = require('./css')
 const sass = require('./sass')
 const moduleCss = require('./module.css')
 const moduleSass = require('./module.sass')
+const nodeModules = require('./node_modules')
 
 module.exports = {
-  babel,
+  file,
   css,
   sass,
   moduleCss,
   moduleSass,
-  file
+  nodeModules,
+  babel
 }