+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 **Please note that Webpacker 3.1.0 and 3.1.1 have some serious bugs so please consider using either 3.0.2 or 3.2.0**
 
-## [Unreleased] - xxxxxx
+## [4.0.0.rc.1] - 2018-12-14
 
 ## Breaking changes